+++ hyperbolix/nn_layers/__init__.py
@@ -13,8 +13,9 @@ from .hyperboloid_activations import (
     hyp_tanh,
 )
 from .hyperboloid_conv import HypConv2DHyperboloid, HypConv3DHyperboloid, LorentzConv2D
-from .hyperboloid_core import hrc, htc
+from .hyperboloid_core import hrc, htc, lorentz_residual
 from .hyperboloid_linear import HTCLinear, HypLinearHyperboloidFHCNN
+from .hyperboloid_positional import HyperbolicRoPE, HypformerPositionalEncoding, hope
 from .hyperboloid_regression import HypRegressionHyperboloid
 from .hyperboloid_regularization import HRCBatchNorm, HRCDropout, HRCLayerNorm
 from .poincare_linear import HypLinearPoincare, HypLinearPoincarePP
@@ -35,7 +36,10 @@ __all__ = [
     "HypRegressionPoincare",
     "HypRegressionPoincareHDRL",
     "HypRegressionPoincarePP",
+    "HyperbolicRoPE",
+    "HypformerPositionalEncoding",
     "LorentzConv2D",
+    "hope",
     "hrc",
     "hrc_gelu",
     "hrc_leaky_relu",
@@ -48,4 +52,5 @@ __all__ = [
     "hyp_relu",
     "hyp_swish",
     "hyp_tanh",
+    "lorentz_residual",
 ]